+++ src/App.js
@@ -1,15 +1,15 @@
 import "./App.css";
 //import { useFeaturedBanners } from "./utils/hooks/useFeaturedBanners";
-import Header from "./Header";
-import Content from "./Content";
-import Footer from "./Footer";
-import { BrowserRouter as Router } from "react-router-dom";
+import Header from "./Components/Header/Header";
+import Content from "./Components/Content/Content";
+import Footer from "./Components//Footer/Footer";
+import { HashRouter as Router } from "react-router-dom"; //using hasrouter since githubpages does not have a configuration to redirect to index.html
 
 function App() {
   //const { data, isLoading } = useFeaturedBanners();
 
   return (
-    <Router basename="/wa-react-2021-q4-internal-capstone-project">
+    <Router>
       <div className="App">
         <Header />
         <Content />
