+++ src/Components/Content/Content.js
@@ -1,19 +1,18 @@
 import "./Content.css";
-import HomePage from "./HomePage";
-import ProductList from "./ProductList";
-import ProductDetails from "./ProductDetails";
+import HomePage from "../Pages/HomePage";
+import ProductList from "../Pages/ProductList";
+import ProductDetails from "../Pages/ProductDetails";
+import SearchResults from "../Pages/SearchResults";
 import { Route, Routes } from "react-router-dom";
 
 const Content = () => {
   return (
     <section className="content">
       <Routes>
-        {["/", "/home"].map((path, index) => (
-          <Route key={index} path={path} element={<HomePage />} />
-        ))}
-
+        <Route path="/" element={<HomePage />} />
         <Route path="/products" element={<ProductList />} />
         <Route path="/products/:productId" element={<ProductDetails />} />
+        <Route path="/search?q={searchTerm}" element={<SearchResults />} />
         <Route path="*" element={<h1>Invalid page!</h1>} />
       </Routes>
     </section>
