+++ src/Components/Header/Header.js
@@ -1,4 +1,4 @@
-import SearchBar from "./SearchBar";
+import SearchBar from "../SearchBar/SearchBar";
 import "./Header.css";
 import { Link } from "react-router-dom";
