--- src/Constants.js
@@ -1,4 +1,0 @@
-export const sitePages = {
-  HOME: "home",
-  ALLPRODUCTS: "allproducts",
-};
